refactor: actions pane header
