+++ src/components/organisms/ActionsPane/ActionsPane.styled.tsx
@@ -2,25 +2,6 @@ import {Button, Skeleton as RawSkeleton, Tabs as RawTabs} from 'antd';
 
 import styled from 'styled-components';
 
-export const Tabs = styled(RawTabs)`
-  width: 100%;
-  height: 100%;
-  overflow: visible;
-
-  & .ant-tabs-nav {
-    padding: 0 16px;
-    margin-bottom: 0px;
-  }
-
-  & .ant-tabs-nav::before {
-    border-bottom: 1px solid #363636;
-  }
-
-  & .ant-tabs-content {
-    height: 100%;
-  }
-`;
-
 export const ActionsPaneContainer = styled.div`
   height: 100%;
   width: 100%;
@@ -52,13 +33,9 @@ export const ActionsPaneMainContainer = styled.div`
   flex-direction: column;
 `;
 
-export const DiffButton = styled(Button)`
-  margin-left: 8px;
-  margin-right: 4px;
-`;
-
-export const SaveButton = styled(Button)`
-  margin-right: 8px;
+export const ExtraRightButton = styled(Button)`
+  padding: 4px 0px;
+  margin-left: 10px;
 `;
 
 export const Skeleton = styled(RawSkeleton)`
@@ -67,15 +44,21 @@ export const Skeleton = styled(RawSkeleton)`
   width: 95%;
 `;
 
-export const LeftArrowButton = styled(Button)`
-  margin-right: 5px;
-`;
+export const Tabs = styled(RawTabs)`
+  width: 100%;
+  height: 100%;
+  overflow: visible;
 
-export const RightArrowButton = styled(Button)`
-  margin-right: 10px;
-`;
+  & .ant-tabs-nav {
+    padding: 0 16px;
+    margin-bottom: 0px;
+  }
 
-export const ExtraRightButton = styled(Button)`
-  padding: 4px 0px;
-  margin-left: 10px;
+  & .ant-tabs-nav::before {
+    border-bottom: 1px solid #363636;
+  }
+
+  & .ant-tabs-content {
+    height: 100%;
+  }
 `;
