+++ src/components/organisms/ActionsPane/ActionsPane.tsx
@@ -4,23 +4,16 @@ import {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import {ReflexContainer, ReflexElement} from 'react-reflex';
 import {useMeasure} from 'react-use';
 
-import {Button, Tabs, Tooltip} from 'antd';
+import {Tabs, Tooltip} from 'antd';
 
-import {ArrowLeftOutlined, ArrowRightOutlined, BookOutlined, CodeOutlined, ContainerOutlined} from '@ant-design/icons';
+import {BookOutlined, CodeOutlined, ContainerOutlined} from '@ant-design/icons';
 
-import {HELM_CHART_ENTRY_FILE, HELM_CHART_HELP_URL, KUSTOMIZE_HELP_URL, TOOLTIP_DELAY} from '@constants/constants';
+import {HELM_CHART_HELP_URL, KUSTOMIZE_HELP_URL, TOOLTIP_DELAY} from '@constants/constants';
 import {makeApplyKustomizationText, makeApplyResourceText} from '@constants/makeApplyText';
 import {
-  ApplyFileTooltip,
-  ApplyTooltip,
-  DiffTooltip,
-  EditPreviewConfigurationTooltip,
-  InstallValuesFileTooltip,
   OpenExternalDocumentationTooltip,
   OpenHelmChartDocumentationTooltip,
   OpenKustomizeDocumentationTooltip,
-  RunPreviewConfigurationTooltip,
-  SaveUnsavedResourceTooltip,
 } from '@constants/tooltips';
 
 import {AlertEnum, AlertType} from '@models/alert';
@@ -29,24 +22,16 @@ import {K8sResource} from '@models/k8sresource';
 
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setAlert} from '@redux/reducers/alert';
-import {openPreviewConfigurationEditor, openResourceDiffModal} from '@redux/reducers/main';
-import {openSaveResourcesToFileFolderModal, setMonacoEditor} from '@redux/reducers/ui';
-import {
-  currentConfigSelector,
-  knownResourceKindsSelector,
-  kubeConfigContextSelector,
-  kubeConfigPathSelector,
-} from '@redux/selectors';
+import {openResourceDiffModal} from '@redux/reducers/main';
+import {setMonacoEditor} from '@redux/reducers/ui';
+import {currentConfigSelector, kubeConfigContextSelector, kubeConfigPathSelector} from '@redux/selectors';
 import {getResourcesForPath} from '@redux/services/fileEntry';
-import {isHelmChartFile, isHelmTemplateFile, isHelmValuesFile} from '@redux/services/helm';
-import {isKustomizationPatch, isKustomizationResource} from '@redux/services/kustomize';
-import {startPreview} from '@redux/services/preview';
-import {isUnsavedResource} from '@redux/services/resource';
+import {isHelmChartFile} from '@redux/services/helm';
+import {isKustomizationResource} from '@redux/services/kustomize';
 import {getResourceSchema, getSchemaForPath, getUiSchemaForPath} from '@redux/services/schema';
 import {applyFileWithConfirm} from '@redux/support/applyFileWithConfirm';
 import {applyHelmChart} from '@redux/thunks/applyHelmChart';
 import {applyResource} from '@redux/thunks/applyResource';
-import {selectFromHistory} from '@redux/thunks/selectionHistory';
 
 import {
   FormEditor,
@@ -54,10 +39,9 @@ import {
   ModalConfirmWithNamespaceSelect,
   Monaco,
   PreviewConfigurationDetails,
-  TitleBar,
 } from '@molecules';
 
-import {Icon, TabHeader} from '@atoms';
+import {TabHeader} from '@atoms';
 
 import WalkThrough from '@components/molecules/WalkThrough';
 
@@ -68,6 +52,7 @@ import {extractFormSchema} from '@src/kindhandlers/common/customObjectKindHandle
 import {getFormSchema, getUiSchema} from '@src/kindhandlers/common/formLoader';
 
 import * as S from './ActionsPane.styled';
+import ActionsPaneHeader from './ActionsPaneHeader';
 
 const {TabPane} = Tabs;
 
@@ -77,17 +62,15 @@ type Props = {
 
 const ActionsPane: React.FC<Props> = ({height}) => {
   const dispatch = useAppDispatch();
-  const applyingResource = useAppSelector(state => state.main.isApplyingResource);
-  const currentSelectionHistoryIndex = useAppSelector(state => state.main.currentSelectionHistoryIndex);
   const fileMap = useAppSelector(state => state.main.fileMap);
   const helmChartMap = useAppSelector(state => state.main.helmChartMap);
   const helmValuesMap = useAppSelector(state => state.main.helmValuesMap);
   const isClusterDiffVisible = useAppSelector(state => state.ui.isClusterDiffVisible);
   const isFolderLoading = useAppSelector(state => state.ui.isFolderLoading);
+  const k8sVersion = useAppSelector(state => state.config.projectConfig?.k8sVersion);
   const kubeConfigContext = useAppSelector(kubeConfigContextSelector);
   const kubeConfigPath = useAppSelector(kubeConfigPathSelector);
   const projectConfig = useAppSelector(currentConfigSelector);
-  const knownResourceKinds = useAppSelector(knownResourceKindsSelector);
   const monacoEditor = useAppSelector(state => state.ui.monacoEditor);
   const paneConfiguration = useAppSelector(state => state.ui.paneConfiguration);
   const previewLoader = useAppSelector(state => state.main.previewLoader);
@@ -96,14 +79,8 @@ const ActionsPane: React.FC<Props> = ({height}) => {
   const selectedPath = useAppSelector(state => state.main.selectedPath);
   const selectedResourceId = useAppSelector(state => state.main.selectedResourceId);
   const selectedValuesFileId = useAppSelector(state => state.main.selectedValuesFileId);
-  const selectionHistory = useAppSelector(state => state.main.selectionHistory);
   const selectedPreviewConfigurationId = useAppSelector(state => state.main.selectedPreviewConfigurationId);
-  const selectedPreviewConfiguration = useAppSelector(state => {
-    if (!selectedPreviewConfigurationId) {
-      return undefined;
-    }
-    return state.config.projectConfig?.helm?.previewConfigurationMap?.[selectedPreviewConfigurationId];
-  });
+  const userDataDir = useAppSelector(state => state.config.userDataDir);
 
   const [activeTabKey, setActiveTabKey] = useState('source');
   const [isApplyModalVisible, setIsApplyModalVisible] = useState(false);
@@ -111,8 +88,6 @@ const ActionsPane: React.FC<Props> = ({height}) => {
   const [isHelmChartApplyModalVisible, setIsHelmChartApplyModalVisible] = useState(false);
   const [selectedResource, setSelectedResource] = useState<K8sResource>();
   const [schemaForSelectedPath, setSchemaForSelectedPath] = useState<any>();
-  const k8sVersion = useAppSelector(state => state.config.projectConfig?.k8sVersion);
-  const userDataDir = useAppSelector(state => state.config.userDataDir);
 
   // Could not get the ref of Tabs Component
   const tabsList = document.getElementsByClassName('ant-tabs-nav-list');
@@ -135,31 +110,9 @@ const ActionsPane: React.FC<Props> = ({height}) => {
     }
   }, [isButtonShrinked, tabsList]);
 
-  const onSaveHandler = () => {
-    if (selectedResource) {
-      dispatch(openSaveResourcesToFileFolderModal([selectedResource.id]));
-    }
-  };
-
   const isKustomization = isKustomizationResource(selectedResource);
   const resourceKindHandler =
     selectedResource && !isKustomization ? getResourceKindHandler(selectedResource.kind) : undefined;
-
-  const isLeftArrowEnabled =
-    selectionHistory.length > 1 &&
-    (currentSelectionHistoryIndex === undefined || (currentSelectionHistoryIndex && currentSelectionHistoryIndex > 0));
-  const isRightArrowEnabled =
-    selectionHistory.length > 1 &&
-    currentSelectionHistoryIndex !== undefined &&
-    currentSelectionHistoryIndex < selectionHistory.length - 1;
-
-  const onClickLeftArrow = () => {
-    selectFromHistory('left', currentSelectionHistoryIndex, selectionHistory, resourceMap, fileMap, dispatch);
-  };
-
-  const onClickRightArrow = () => {
-    selectFromHistory('right', currentSelectionHistoryIndex, selectionHistory, resourceMap, fileMap, dispatch);
-  };
 
   const applySelection = useCallback(() => {
     if (selectedValuesFileId && (!selectedResourceId || selectedValuesFileId === selectedResourceId)) {
@@ -217,19 +170,6 @@ const ActionsPane: React.FC<Props> = ({height}) => {
     },
     [dispatch]
   );
-
-  const isDiffButtonDisabled = useMemo(() => {
-    if (!selectedResource) {
-      return true;
-    }
-    if (isKustomizationPatch(selectedResource) || isKustomizationResource(selectedResource)) {
-      return true;
-    }
-    if (!knownResourceKinds.includes(selectedResource.kind)) {
-      return true;
-    }
-    return false;
-  }, [selectedResource, knownResourceKinds]);
 
   const onClickApplyResource = useCallback(
     (namespace?: {name: string; new: boolean}) => {
@@ -329,13 +269,6 @@ const ActionsPane: React.FC<Props> = ({height}) => {
     }
   }, [selectedResource, activeTabKey, resourceKindHandler, isKustomization, selectedPath, schemaForSelectedPath]);
 
-  const isSelectedResourceUnsaved = useCallback(() => {
-    if (!selectedResource) {
-      return false;
-    }
-    return isUnsavedResource(selectedResource);
-  }, [selectedResource]);
-
   useEffect(() => {
     if (tabsList && tabsList.length && extraButton.current) {
       getDistanceBetweenTwoComponents();
@@ -346,112 +279,13 @@ const ActionsPane: React.FC<Props> = ({height}) => {
     setSchemaForSelectedPath(selectedPath ? getSchemaForPath(selectedPath, fileMap) : undefined);
   }, [selectedPath, fileMap]);
 
-  const deployTooltip = useMemo(() => {
-    return selectedPath ? (isHelmValuesFile(selectedPath) ? InstallValuesFileTooltip : ApplyFileTooltip) : ApplyTooltip;
-  }, [selectedPath]);
-
-  const isDeployButtonDisabled = useMemo(() => {
-    return (
-      (!selectedResourceId && !selectedPath) ||
-      (selectedPath && selectedPath.endsWith(HELM_CHART_ENTRY_FILE)) ||
-      (selectedPath && isHelmTemplateFile(selectedPath)) ||
-      (selectedResource &&
-        !isKustomizationResource(selectedResource) &&
-        (isKustomizationPatch(selectedResource) || !knownResourceKinds.includes(selectedResource.kind)))
-    );
-  }, [selectedResource, knownResourceKinds, selectedResourceId, selectedPath]);
-
-  const onClickRunPreviewConfiguration = useCallback(() => {
-    if (!selectedPreviewConfiguration) {
-      return;
-    }
-    startPreview(selectedPreviewConfiguration.id, 'helm-preview-config', dispatch);
-  }, [dispatch, selectedPreviewConfiguration]);
-
-  const onClickEditPreviewConfiguration = useCallback(() => {
-    if (!selectedPreviewConfiguration) {
-      return;
-    }
-    const chart = Object.values(helmChartMap).find(c => c.filePath === selectedPreviewConfiguration.helmChartFilePath);
-    if (!chart) {
-      return;
-    }
-    dispatch(
-      openPreviewConfigurationEditor({
-        helmChartId: chart.id,
-        previewConfigurationId: selectedPreviewConfiguration.id,
-      })
-    );
-  }, [dispatch, selectedPreviewConfiguration, helmChartMap]);
-
   return (
     <S.ActionsPaneMainContainer ref={actionsPaneRef}>
-      {selectedPreviewConfigurationId ? (
-        <TitleBar title="Helm Command">
-          <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title={RunPreviewConfigurationTooltip} placement="bottomLeft">
-            <Button type="primary" size="small" ghost onClick={onClickRunPreviewConfiguration}>
-              Preview
-            </Button>
-          </Tooltip>
-          <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title={EditPreviewConfigurationTooltip} placement="bottomLeft">
-            <S.DiffButton size="small" type="primary" ghost onClick={onClickEditPreviewConfiguration}>
-              Edit
-            </S.DiffButton>
-          </Tooltip>
-        </TitleBar>
-      ) : (
-        <TitleBar title="Editor">
-          <>
-            <S.LeftArrowButton
-              onClick={onClickLeftArrow}
-              disabled={!isLeftArrowEnabled}
-              type="link"
-              size="small"
-              icon={<ArrowLeftOutlined />}
-            />
-            <S.RightArrowButton
-              onClick={onClickRightArrow}
-              disabled={!isRightArrowEnabled}
-              type="link"
-              size="small"
-              icon={<ArrowRightOutlined />}
-            />
-
-            {isSelectedResourceUnsaved() && (
-              <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title={SaveUnsavedResourceTooltip}>
-                <S.SaveButton id="save-button" type="primary" size="small" onClick={onSaveHandler}>
-                  Save
-                </S.SaveButton>
-              </Tooltip>
-            )}
-
-            <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title={deployTooltip} placement="bottomLeft">
-              <Button
-                loading={Boolean(applyingResource)}
-                type="primary"
-                size="small"
-                ghost
-                onClick={applySelection}
-                disabled={isDeployButtonDisabled}
-                icon={<Icon name="kubernetes" />}
-              >
-                {selectedPath && isHelmValuesFile(selectedPath) ? 'Install' : 'Deploy'}
-              </Button>
-            </Tooltip>
-            <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title={DiffTooltip} placement="bottomLeft">
-              <S.DiffButton
-                size="small"
-                type="primary"
-                ghost
-                onClick={diffSelectedResource}
-                disabled={isDiffButtonDisabled}
-              >
-                Diff
-              </S.DiffButton>
-            </Tooltip>
-          </>
-        </TitleBar>
-      )}
+      <ActionsPaneHeader
+        selectedResource={selectedResource}
+        applySelection={applySelection}
+        diffSelectedResource={diffSelectedResource}
+      />
 
       <ReflexContainer orientation="horizontal" style={{height: height - 40}}>
         <ReflexElement flex={1.0}>
