+++ src/components/organisms/ActionsPane/ActionsPaneHeader.styled.tsx
@@ -1,0 +1,20 @@
+import {Button} from 'antd';
+
+import styled from 'styled-components';
+
+export const DiffButton = styled(Button)`
+  margin-left: 8px;
+  margin-right: 4px;
+`;
+
+export const LeftArrowButton = styled(Button)`
+  margin-right: 5px;
+`;
+
+export const RightArrowButton = styled(Button)`
+  margin-right: 10px;
+`;
+
+export const SaveButton = styled(Button)`
+  margin-right: 8px;
+`;
